feat: added XML documentation and improved Exception handling
