+++ EntityMigration/Exceptions/InvalidMigrationException.cs
@@ -1,7 +1,13 @@
 namespace EntityMigration.Exceptions;
 
-public class InvalidMigrationException : Exception
+/// <summary>
+/// Exception thrown when invalid migration configuration is detected
+/// </summary>
+public class InvalidMigrationException : MigrationException
 {
+    /// <summary>
+    /// Initializes a new instance with a message
+    /// </summary>
     public InvalidMigrationException(string message)
         : base(message) { }
 }
