+++ EntityMigration/Exceptions/MigrationPathNotFoundException.cs
@@ -1,7 +1,13 @@
 namespace EntityMigration.Exceptions;
 
-public class MigrationPathNotFoundException : Exception
+/// <summary>
+/// Exception thrown when no valid migration path exists between requested types
+/// </summary>
+public class MigrationPathNotFoundException : MigrationException
 {
+    /// <summary>
+    /// Initializes a new instance with a message
+    /// </summary>
     public MigrationPathNotFoundException(string message)
         : base(message) { }
 }
